+++ lib/mix/shopifex/migration.ex
@@ -198,4 +198,7 @@ defmodule Mix.Shopifex.Migration do
 
   defp to_migration_index(table, {key_or_keys, :gin}),
     do: "create index(:#{table}, #{inspect(List.wrap(key_or_keys))}, using: \"GIN\")"
+
+  defp to_migration_index(table, {key_or_keys, :index}),
+    do: "create index(:#{table}, #{inspect(List.wrap(key_or_keys))})"
 end